Move Authenticode machinery into an Authenticode subfolder
Group the Authenticode signature reading, PE hashing and trust verification
types under an Authenticode subfolder and namespace. The validators and their
options remain in the project root.
