+++ src/PluginSystem/SAF.PluginSystem.Hosting.Extensions/Authenticode/WindowsAuthenticodeTrustVerifier.cs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: MPL-2.0
 
-namespace SAF.PluginSystem.Hosting.Extensions;
+namespace SAF.PluginSystem.Hosting.Extensions.Authenticode;
 
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
